Try to use MVC, can add and remove users, some refactoring...
