+++ src/main/java/su/kotindustries/kotnotes/controllers/MainController.java
@@ -1,0 +1,14 @@
+package su.kotindustries.KotNotes.controllers;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class MainController {
+
+    @GetMapping("/")
+    public String getMain(Model model){
+        return "MainView";
+    }
+}
