+++ src/main/resources/templates/MainView.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <title>KotNotes</title>
+    </head>
+    <body>
+        <h2>Main page</h2>
+        Sample text
+    <a href='/users'>User Management</a>
+    </body>
+</html>
